TYPO: renaming obj to obs in file name
